Implement the jwt-bearer grant type according to RFC7523 (issue #260) This will allow resource servers to obtain tokens for requests to other resource servers further upstream, which may need scopes that are not in the current token.
The grant_type parameter must
"urn:ietf:params:oauth:grant-type:jwt-bearer".

This is similar to password grant, but instead of taking a username and
password, a token containing a user_id is used instead. This token is
not submitted in the header (that's already used by the client's basic
auth) but POSTed in the form parameter "assertion".

Clients must be authorized to use this grant type, and like password
grant, non-admins cannot authorize this grant type to clients.

The identity token (could be an id_token or a user's access_token) aud
claim must contain the client_id that is submitting the token request.

Example token request with curl, where resource1 is requesting a token
with scope resource2.read, because the current token doens't contain
that scope:

$ curl http://resource1:secret@localhost:8080/uaa/oauth/token \
 --data-urlencode
'grant_type=urn:ietf:params:oauth:grant-type:jwt-bearer' \
--data-urlencode 'assertion=[resource1's current token, containing user
identity] \
--data-urlencode 'scope=resource2.read' \